Force use of Hugo embedded link render hook
The default Hugo link render hook seems to pass through the link
destination largely unmodified, contrary to the documentation which
states that it will be treated as relative to the baseURL.  To obtain
the documented behaviour, it seems to be necessary to explicitly
enable the non-default "embedded" link render hook.

The syntax for enabling the embedded link render hook changed in Hugo
version 0.148.  Include both versions (using "enableDefault" and using
"useEmbedded") to allow for interoperability with both.

Using the embedded link render hook also allows the links within the
Markdown text to include the ".md" file suffix, which allows for
easier traversal of files within the editor.  Standardise on this to
simplify the editing process.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>