Use page bundles
Move each page to a page bundle, to allow for e.g. per-page header
background images or social media preview images.

For the home page, create a dummy _index.md to prevent build warnings
and specify the root URL explicitly in home/index.md.  This causes the
home page bundle to render to the correct destination, with the caveat
that .IsHome and related functions will not work.

Signed-off-by: Michael Brown <michael@bar-oh.com>