Parse Markdown attributes from link titles
Hugo allows for Markdown attributes on block elements and titles, but
does not yet support the equivalent syntax on links:

  [Launch thing](https://launch.thing) { .button target="_blank" }

Work around this limitation by parsing (and stripping) a Markdown
attribute string from the link title, allowing for the very similar
looking:

  [Launch thing](https://launch.thing "{ .button target=\"_blank\" }")

There is no function available within Hugo for parsing a Markdown
attribute string, so we parse this manually using regular expressions,
allowing for most patterns that we expect to use.

With this feature in place, the dedicated shortcode for generating a
link button for the Ripples drink printer can be removed and replaced
with a standard link.

Signed-off-by: Michael Brown <michael@bar-oh.com>