Use colour vector CSS variables to reduce repetition of colours
The pulsing box shadow effect requires the animation keyframe to be
defined using the same colour at two different alpha values.  This is
not a calculation that can be performed in native CSS.

Work around the limitation by using colour vector values (i.e. raw RGB
tuples) in CSS variables, which can then be used in rgba() expressions
to construct the same colour with variable transparency.

Define an SCSS function rgbv() to extract the RGB vector from a
colour, so that colour names can be used and so that colour
highlighting within the editor works as expected.

Signed-off-by: Michael Brown <michael@bar-oh.com>