Enlarge Google Map vertically to improve scroll target behaviour
The Google Map is used as the anchor target for the "Find us"
nagivation link.  Since the map is at the end of the home page, this
can cause the link to appear broken if the page is already scrolled to
the end, since no further scroll is possible.

Set the map height to 75% of the viewport, thereby ensuring that it
always acts as a sensible anchor target.

Since the map lies below the booking form and opening hours table,
this fix also ensures that both of those will similarly act as
sensible anchor targets.

Signed-off-by: Michael Brown <michael@bar-oh.com>