Show a nav link in header (rather than a nav label) while nav is open
If the nav bar has been opened and the user has scrolled the page,
this will re-expose the nav label at the top of the header.  Clicking
on this will currently toggle the nav back to being closed, which is
unlikely to match user expectations.

While the nav is open, swap the label in the header for a link that
will scroll back up to the top of the nav, thereby matching the most
likely user expectation.

Signed-off-by: Michael Brown <michael@bar-oh.com>